+++ NEWS.md
@@ -1,6 +1,15 @@
+- Version 0.10.0 - released 2020-08-02
+  - Added support for arbitrary clipboard commands - thanks @ign0tus!
+  - `--clipboard` option deprecated in favour of `--clipboard-cmd`
+  - Added support for arbitrary menu command via `--menu-cmd`
+  - `--dmenu-prompt` option and trailing dmenu arguments deprecated in favour of
+    `--menu-cmd`
+  - Fixed bug where `wtype` would be preferred over `xdotool` if installed, even
+    if current environment is X11 and not Wayland - thanks Mark Stosberg!
+
 - Version 0.9.0 - released 2020-07-22
   - If `wtype` is available, it will be preferred over `xdotool` when `--type`
-    is used
+    is used - thanks Mark Stosberg!
 
 - Version 0.8.0 - released 2019-04-10
   - Will now prompt for YubiKey OATH password if needed